Rename the event-gateway component to event-service
