--- components/event-gateway/Dockerfile
@@ -1,19 +1,0 @@
-FROM golang:1.9-alpine as builder
-
-ARG DOCK_PKG_DIR=/go/src/github.com/kyma-project/kyma/components/event-gateway
-
-RUN mkdir -p $DOCK_PKG_DIR
-
-COPY ./ $DOCK_PKG_DIR
-WORKDIR $DOCK_PKG_DIR
-
-RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o eventgateway ./cmd/eventgateway
-
-FROM alpine
-RUN apk update && apk add curl ngrep
-
-LABEL source=git@github.com:kyma-project/kyma.git
-
-COPY --from=builder /go/src/github.com/kyma-project/kyma/components/event-gateway .
-
-CMD ["/eventgateway"]
