+++ components/event-service/Gopkg.toml
