+++ components/event-service/internal/events/bus/send.go
@@ -7,9 +7,9 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/kyma-project/kyma/components/event-gateway/internal/events/api"
-	"github.com/kyma-project/kyma/components/event-gateway/internal/httpconsts"
-	"github.com/kyma-project/kyma/components/event-gateway/internal/httptools"
+	"github.com/kyma-project/kyma/components/event-service/internal/events/api"
+	"github.com/kyma-project/kyma/components/event-service/internal/httpconsts"
+	"github.com/kyma-project/kyma/components/event-service/internal/httptools"
 )
 
 var (
