+++ components/event-service/internal/events/shared/constants.go
