+++ components/event-service/internal/externalapi/healthcheckhandler.go
