+++ components/event-service/internal/externalapi/healthcheckhandler_test.go
