+++ components/event-service/internal/httpconsts/httpconsts.go
