+++ components/event-service/internal/httperrors/httperrors.go
